Removed Director's gametic
This is the beginning of fixing https://github.com/alelazar/galcon/issues/3 as gametic was used in several areas of the codebase. However, it turns out all these areas stemmed from the fact that Planet used it for timing production. I changed Planets to keep track of gametic internally and made all areas that used gametic reference it from the planet in question instead. 

I left "tic" itself alone as it will be necessary in the future for doing proper logging. 

I also changed update back to being a void overridden from Unit and made a few formatting changes.