Removed stats (#26)
This resolves https://github.com/alelazar/galcon/issues/24. 

The Stats system was well intentioned but is quite clunky. It will be better to have stats be a separate thing that parses game logs, as this also allows for multiple stats on the same game and stats on historical games. As such I am removing it entirely for the time being.