Removed Main, rolled it into Director (#37)
This moves towards making https://github.com/alelazar/galcon/issues/3 better but does not yet solve it. 

This change allows for removal of the clunky static calls to Main and Director. It also cleans things up a little in general and allows for other refactors to be more easily implemented.