GameSettings refactor (#45)
Changed GameSettings to read from a file rather than need to be directly set. 

Removed GameSettings as a super class from where it was being used as such. 

Did a few other random refactors.

This design is not final but just done to get GameSettings to be reading from a file. Further exploration needs to be done for whether to use a singleton or pass around the game settings. 

This makes progress on https://github.com/alelazar/galcon/issues/15 and fixes https://github.com/alelazar/galcon/issues/14